css first idea
got loads of css yesterday added to my portfolio, it was fun overall. while adding css I realize somethings in html were uncalled or too complicated to goi ahead with what I had in mind. Got a reset.css to reset UL and A with no decorations in my nav menu or along in my social media links, also reset all images to 100% whatsover and html margin, padding, borders and outline to 0 or none. Im still not happy with my profile picture so it might change. In the style css I used transitions, hovers and scroll-behavior to give the user a better experience. Along the website display Flex and its properties were widely used for styling as I don't think Grid would match as nicely for a Portfolio.