css styling + responsive.css creation
created a filed called responsive.css for screens with max width of 1280px and 425px using media queries so my layout will adapt. I'm still working on the footer, but got a logo ready and positioned on top and also finish the section projects, will add on one new html file for each project. Also got some animation with an arrow down and a gradient background.